refactor : 랜덤 번호 생성 테스트하기 쉬운 코드로 수정
