+++ README(step2).md
@@ -11,6 +11,9 @@
     - [x] Input값 parsing 하는 객체에 분산된 parsing책임을 전부 부여하기
     - [x] 구매로직으로부터 구매개수 구하는 로직 분리하기
   
+  - 테스트
+    - [x] 랜덤 번호 생성 테스트하기 쉬운 코드로 수정
+  
   - 번외
     - [x] Scanner 자원 해제하도록 코드 수정
     - [x] 단 하나의 인스턴스만 필요한 클래스 싱글톤 패턴으로 수정하기
