+++ src/main/java/LottoConfig.java
@@ -1,6 +1,6 @@
 import controller.LottoController;
 import controller.dto.LottoDtoMapper;
-import model.LottoNumberGenerator;
+import model.LottoNumbersGenerator;
 import model.LottoRankCalculator;
 import model.LottoStore;
 import view.LottoParser;
@@ -28,6 +28,6 @@ public final class LottoConfig {
     }
 
     private LottoStore lottoStore() {
-        return new LottoStore(new LottoNumberGenerator(), new LottoRankCalculator());
+        return new LottoStore(new LottoNumbersGenerator(), new LottoRankCalculator());
     }
 }
