--- src/main/java/common/NumberGenerator.java
@@ -1,16 +1,0 @@
-package common;
-
-import java.util.List;
-import java.util.Random;
-
-public class NumberGenerator {
-
-    public static List<Integer> generate(int size, int minNumber, int maxNumber) {
-        return new Random()
-                .ints(minNumber, maxNumber + 1)
-                .distinct()
-                .limit(size)
-                .boxed()
-                .toList();
-    }
-}
