+++ src/main/java/common/NumbersGenerator.java
@@ -1,0 +1,8 @@
+package common;
+
+import java.util.List;
+
+public interface NumbersGenerator {
+
+    List<Integer> generate();
+}
